Move consts, use named import
