+++ pkg/hints/math_cmp_hints.go
@@ -1,7 +1,7 @@
 package hints
 
 import (
-	. "github.com/lambdaclass/cairo-vm.go/pkg/builtins"
+	"github.com/lambdaclass/cairo-vm.go/pkg/builtins"
 	. "github.com/lambdaclass/cairo-vm.go/pkg/hints/hint_utils"
 	. "github.com/lambdaclass/cairo-vm.go/pkg/lambdaworks"
 	. "github.com/lambdaclass/cairo-vm.go/pkg/vm"
@@ -14,7 +14,7 @@ func isNN(ids IdsManager, vm *VirtualMachine) error {
 	if err != nil {
 		return err
 	}
-	if a.Bits() < RANGE_CHECK_N_PARTS*INNER_RC_BOUND_SHIFT {
+	if a.Bits() < builtins.RANGE_CHECK_N_PARTS*builtins.INNER_RC_BOUND_SHIFT {
 		return vm.Segments.Memory.Insert(vm.RunContext.Ap, NewMaybeRelocatableFelt(FeltZero()))
 	}
 	return vm.Segments.Memory.Insert(vm.RunContext.Ap, NewMaybeRelocatableFelt(FeltOne()))
@@ -26,7 +26,7 @@ func isNNOutOfRange(ids IdsManager, vm *VirtualMachine) error {
 	if err != nil {
 		return err
 	}
-	if (FeltZero().Sub(a).Sub(FeltOne())).Bits() < RANGE_CHECK_N_PARTS*INNER_RC_BOUND_SHIFT {
+	if (FeltZero().Sub(a).Sub(FeltOne())).Bits() < builtins.RANGE_CHECK_N_PARTS*builtins.INNER_RC_BOUND_SHIFT {
 		return vm.Segments.Memory.Insert(vm.RunContext.Ap, NewMaybeRelocatableFelt(FeltZero()))
 	}
 	return vm.Segments.Memory.Insert(vm.RunContext.Ap, NewMaybeRelocatableFelt(FeltOne()))
